fix all the bugs? as far as I can tell? also move rehash to forest, not undo file
