+++ utreexo/forestproofs.go
@@ -145,9 +145,9 @@ func (f *Forest) ProveBlock(hs []Hash) (BlockProof, error) {
 		return bp, nil
 	}
 
-	for h, p := range f.positionMap {
-		fmt.Printf("%x@%d ", h[:4], p)
-	}
+	// for h, p := range f.positionMap {
+	// fmt.Printf("%x@%d ", h[:4], p)
+	// }
 
 	// first get all the leaf positions
 	// there shouldn't be any duplicates in hs, but if there are I guess
