+++ utreexo/undo.go
@@ -133,103 +133,3 @@ func (f *Forest) BuildUndoData(numadds uint64, dels []uint64) *undoBlock {
 
 	return ub
 }
-
-// reHash hashes new data in the forest based on dirty positions.
-// right now it seems "dirty" means the node itself moved, not that the
-// parent has changed children.
-// TODO: switch the meaning of "dirt" to mean parents with changed children;
-// this will probably make it a lot simpler.
-func (f *Forest) reHash(dirt []uint64) error {
-	if f.height == 0 || len(dirt) == 0 { // nothing to hash
-		return nil
-	}
-	tops, topheights := getTopsReverse(f.numLeaves, f.height)
-	fmt.Printf("nl %d f.h %d tops %v\n", f.numLeaves, f.height, tops)
-
-	dirty2d := make([][]uint64, f.height)
-	h := uint8(0)
-	dirtyRemaining := 0
-	for _, pos := range dirt {
-		if pos > f.numLeaves {
-			return fmt.Errorf("Dirt %d exceeds numleaves %d", pos, f.numLeaves)
-		}
-		dHeight := detectHeight(pos, f.height)
-		// increase height if needed
-		for h < dHeight {
-			h++
-		}
-		if h > f.height {
-			return fmt.Errorf("postion %d at height %d but forest only %d high",
-				pos, h, f.height)
-		}
-		// if bridgeVerbose {
-		// fmt.Printf("h %d\n", h)
-		// }
-		dirty2d[h] = append(dirty2d[h], pos)
-		dirtyRemaining++
-	}
-
-	// this is basically the same as VerifyBlockProof.  Could maybe split
-	// it to a separate function to reduce redundant code..?
-	// nah but pretty different beacuse the dirtyMap has stuff that appears
-	// halfway up...
-
-	var currentRow, nextRow []uint64
-
-	// floor by floor
-	for h = uint8(0); h < f.height; h++ {
-		if bridgeVerbose {
-			fmt.Printf("dirty %v\ncurrentRow %v\n", dirty2d[h], currentRow)
-		}
-
-		// merge nextRow and the dirtySlice.  They're both sorted so this
-		// should be quick.  Seems like a CS class kindof algo but who knows.
-		// Should be O(n) anyway.
-
-		currentRow = mergeSortedSlices(currentRow, dirty2d[h])
-		dirtyRemaining -= len(dirty2d[h])
-		if dirtyRemaining == 0 && len(currentRow) == 0 {
-			// done hashing early
-			break
-		}
-
-		for i, pos := range currentRow {
-			// skip if next is sibling
-			if i+1 < len(currentRow) && currentRow[i]|1 == currentRow[i+1] {
-				continue
-			}
-			if len(tops) == 0 {
-				return fmt.Errorf(
-					"currentRow %v no tops remaining, this shouldn't happen",
-					currentRow)
-			}
-			// also skip if this is a top
-			if pos == tops[0] {
-				continue
-			}
-
-			right := pos | 1
-			left := right ^ 1
-			parpos := up1(left, f.height)
-
-			//				fmt.Printf("bridge hash %d %04x, %d %04x -> %d\n",
-			//					left, leftHash[:4], right, rightHash[:4], parpos)
-			if f.forest[left] == empty || f.forest[right] == empty {
-				f.forest[parpos] = empty
-			} else {
-				par := Parent(f.forest[left], f.forest[right])
-				f.HistoricHashes++
-				f.forest[parpos] = par
-			}
-			nextRow = append(nextRow, parpos)
-		}
-		if topheights[0] == h {
-			tops = tops[1:]
-			topheights = topheights[1:]
-		}
-		currentRow = nextRow
-		nextRow = []uint64{}
-	}
-
-	return nil
-}
